+++ R/zzz.R
@@ -18,14 +18,10 @@ utils::globalVariables(c(".", "%>%", ":="))
   packageStartupMessage(message)
 }
 
-#' @importFrom magrittr %>%
 #' @importFrom data.table := %between% fifelse %chin%
 #' @importFrom stats na.omit
 #' @importFrom utils head tail object.size
 #' @importFrom Rcpp compileAttributes
-#' @importFrom lwgeom st_geod_length
-#' @importFrom readr write_rds
-#' @importFrom pbapply pblapply
 #' @useDynLib gtfs2gps, .registration = TRUE
 NULL