+++ NEWS.md
@@ -4,7 +4,11 @@
 * Major changes
   * Filter functions were removed from the package because gtftools alredy implements them
   * Function `gtfs2gps` now uses `parallel = TRUE` by default.
-  * Function `gtfs2gps` now uses `gtfstools::frequencies_to_stop_times()` to convert frequency-based GTFS feeds to stop times before processing the data
+  * Function `gtfs2gps` now uses `gtfstools::frequencies_to_stop_times()` to convert. frequency-based GTFS feeds to stop times before processing the data.
+
+* Minor changes
+  * Removed `readr`, `pbapply`, `lwgeom` and `magrittr` from package dependencies.
+
 
 # gtfs2gps v2.0-3
 * Minor changes